// src_old/interfaces.ts
export interface Product {
  id?: string;
  lotNo: string;
  denier: string;
  type: string;
  noOfFilaments: string;
  luster: string;
  userId: string;
}

export type ProductOrderType = Omit<Product, 'id'> & {
  id: number;
  quantity: number;
  rate: number;
  endUses: EndUseWithColor[];
  [key: string]: any;
};

export interface ProductOrder {
  id: number;
  quantity: number;
  rate: number;
  product: ProductData;
}

export interface ProductData extends Product {
  endUses: EndUse[];
}

export interface OrderUpdateData {
  id: string;
  quantity: number;
  rate: number;
  product: ProductData;
  salesReceiptDate?: string;
  validity?: string;
  orderProducts?: {
    productId: number;
    quantity: number;
    rate: number;
  }[];
  shipmentType?: string;
  PI_number?: string;
  note?:string;
  paymentTypeId?: number;
  currencyUnitId?: number;
  customerId?: number;
}

export interface ProductUpdateData extends Product {
  endUses: number[];
}
export interface Address {
  id?: number;
  street: string;
  city: string;
  state: string;
  country: string;
  postalCode: string;
}

export interface EndUse {
  id?: number;
  name: string;
  description?: string;
}

export interface EndUseWithColor extends EndUse {
  color: string;
}

export interface EndUsesListProps {
  endUses: EndUseWithColor[];
}

export interface Contact {
  id: number;
  name: string;
  designation: string;
  contactNumber: string;
  address?: Address;
  billingAddress: string;
  isMainContact?: boolean;
}

export interface Customer {
  id: number;
  businessName: string;
  contacts: Contact[];
  mainContact?: Contact;
  [key: string]: any;
}
// ...
export type CurrencyUnit = {
  id: number;
  name: string;
  code: string;
  symbol: string;
};

export type PaymentMethod = {
  id: number;
  name: string;
};

export interface Order {
  id: number;
  salesReceiptDate: string;
  validity: string;
  shipmentType: string;
  PI_number: string;
  note:string;
  paymentType: PaymentMethod;
  currencyUnit: CurrencyUnit;
  customer: Customer;
  orderProducts: ProductOrder[];
  [key: string]: any;
}